Details View UI integration compeleted

+++ VO2Video/EngineModule/Models/Movies/MovieDetailsModel.swift
@@ -8,7 +8,8 @@
 import Foundation
 
 struct MovieDetails: Codable, MovieProtocol {
-    let id: Int64
+    
+    let id: Int
     let overview: String?
     let rating: Double?
     let title: String?
